create separate component for house as a item from list
