+++ src/components/HomePage/House/House.module.scss
@@ -1,0 +1,90 @@
+@import '/src/theme/colors.scss';
+@import '/src/theme/breakpoint.scss';
+
+.houseElement {
+  border: 1px solid #d6d6d6;
+  border-radius: 4px;
+  margin-bottom: 7px;
+  position: relative;
+  height: fit-content;
+
+  h4 {
+    margin: 0;
+    padding: 10px 0 0 10px;
+    font-size: 20px;
+    line-height: 24px;
+    letter-spacing: 0.15px;
+    font-family: Roboto;
+    color: $text-color;
+  }
+
+  img {
+    object-fit: cover;
+    width: 100%;
+    height: 240px;
+  }
+}
+
+.price {
+  font-family: Roboto;
+  margin: 0;
+  padding: 5px 0 10px 10px;
+  color: $text-color;
+  font-size: 12px;
+  line-height: 17px;
+}
+
+.shortInfo {
+  margin: 0;
+  padding: 16px 0 0 5px;
+  text-align: left;
+  font-size: 14px;
+  line-height: 20px;
+  font-family: Roboto;
+  letter-spacing: 0.25px;
+  color: #00000099;
+}
+
+.moreInfo {
+  padding: 0;
+  margin: 22px 0 29px 10px;
+  color: $turquoise-color;
+}
+
+.moreInfo:hover {
+  background-color: $white-color;
+}
+
+.icon {
+  position: absolute !important;
+  right: 10px;
+  padding: 0 !important;
+  padding-top: 7px !important;
+  text-align: right;
+  background-color: $white-color;
+}
+
+@media screen and (min-width: $screen-xl) {
+  .houseElement {
+    margin-right: 0px;
+  }
+}
+
+@media screen and (min-width: $screen-md) {
+  .houseElement {
+    margin-right: 5px;
+  }
+}
+
+@media screen and (min-width: $screen-sm) {
+  .houseElement {
+    margin-right: 5px;
+    width: 370px;
+  }
+}
+
+@media screen and (min-width: $screen-xs) {
+  .houseElement {
+    width: 290px;
+  }
+}
