+++ src/components/HomePage/ListOfHouses/ListOfHouses.tsx
@@ -1,17 +1,14 @@
 /* eslint-disable react/jsx-props-no-spreading */
 /* eslint-disable @typescript-eslint/no-redeclare */
 import DoneIcon from '@mui/icons-material/Done';
-import FavoriteIcon from '@mui/icons-material/Favorite';
-import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
-import { Autocomplete, Box, Button, Checkbox, TextField } from '@mui/material';
+import House from '../House/House';
+import { Autocomplete, Box, TextField } from '@mui/material';
 import Chip from '@mui/material/Chip';
 import { useEffect, useState } from 'react';
-import noPhoto from 'src/assets/images/nophoto.png';
 
 import styles from './ListOfHouses.module.scss';
 
 const options = ['Payment (Low to High)', 'Payment (High to Low)', 'A-Z', 'Z-A'] as const;
-const label: { inputProps: { 'aria-label': string } } = { inputProps: { 'aria-label': 'Checkbox demo' } };
 
 function ListOfHouses({ houses }: { houses: BasicHouseData[] }) {
   const [sortType, setSortType] = useState<string | null>(null);
@@ -69,24 +66,7 @@ function ListOfHouses({ houses }: { houses: BasicHouseData[] }) {
       </Box>
       <Box component="div" className={styles.housesList}>
         {sortedHouses.map((item: BasicHouseData, i) => (
-          <Box component="div" className={styles.houseElement} key={i.toString()}>
-            <h4>
-              {item.city}, {item.streetName} {item.streetNumber}
-            </h4>
-            <p className={styles.price}>
-              {item.price}zł/m<sup>2</sup>
-            </p>
-            <img src={item.photo_0 ? item.photo_0 : noPhoto} alt="House" />
-            <p className={styles.shortInfo}>{item.descriptionField}</p>
-            <Button className={styles.moreInfo}>more info</Button>
-            <Checkbox
-              color="warning"
-              {...label}
-              icon={<FavoriteBorderIcon />}
-              checkedIcon={<FavoriteIcon />}
-              className={styles.icon}
-            />
-          </Box>
+          <House house={item} key={i} />
         ))}
       </Box>
     </Box>
